% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey}
\alias{survey}
\title{STRIPS Survey data}
\format{A data.frame with 18 variables the first of which is \code{CaseID} 
and the remainder are survey questions.}
\usage{
survey
}
\description{
STRIPS Survey data for questions 3 and 15
}
\examples{
survey
}
\keyword{datasets}

