% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey_questions}
\alias{survey_questions}
\title{STRIPS Survey questions}
\format{A data.frame with a 2 variables: \code{question} and 
\code{question_text}.}
\usage{
survey_questions
}
\description{
To aid analysis, survey data are recorded using the question number and this
data.frame provides the associated wording for the question.
}
\author{
J. Gordon Arbuckle, \email{arbuckle@iastate.edu}
}
\seealso{
\link{\code{survey_responses}}
}
\keyword{datasets}

