% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey_responses}
\alias{survey_responses}
\title{STRIPS Survey data}
\format{A data.frame with a lot of variables the first of which is 
\code{CaseID} and the remaining 123 columns are survey questions.}
\usage{
survey_responses
}
\description{
STRIPS Survey data for all questions. At the moment, this data set does not 
read in NA's properly and therefore caution should be used when looking at 
questions other than Q3 and Q15
}
\author{
J. Gordon Arbuckle, \email{arbuckle@iastate.edu}
}
\seealso{
\link{\code{survey_questions}}
}
\keyword{datasets}

